share code for displaying staked amounts helper
